% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge.R
\name{badge}
\alias{badge}
\title{Create a badge}
\usage{
badge(ref = NULL)
}
\arguments{
\item{ref}{Reference for a GitHub repository. If \code{NULL} (the default),
the reference is determined by the URL field in the DESCRIPTION file.}
}
\description{
Create a badge showing the status of your R package in 
the \code{rworkflows} GitHub action.
Uses the package 
\href{https://github.com/GuangchuangYu/badger}{\pkg{badger}}.
}
\examples{
## Causes issues bc examples can't find the  the DESCRIPTION file.
 \dontrun{
 badger::badge_github_actions() 
 }
}
