% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{biocpkgtools_db}
\alias{biocpkgtools_db}
\title{Static Bioc packages list}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 100 rows and 53 columns.
}
\source{
\code{
as_ascii <- function(dt,
                     cols=names(dt)){
  cols <- cols[cols %in% names(dt)]
  func <- function(v){
    Encoding(v) <- "latin1"
    iconv(v, "latin1", "UTF-8")
  }
  for(col in cols){
    if(is.character(dt[[col]])){
      dt[[col]] <- func(dt[[col]])
    }
  }
  return(dt)
}
biocpkgtools_db <- get_description_repo_biocpkgtools(repo="BioCsoft")
biocpkgtools_db <-  as_ascii(biocpkgtools_db[seq(100)])
usethis::use_data(biocpkgtools_db, overwrite = TRUE)
}
}
\usage{
data("biocpkgtools_db")
}
\description{
A static snapshot of all Bioconductor packages from 
\link[BiocPkgTools]{biocPkgList}.
Last updated: Sept. 06 2023
}
\keyword{datasets}
