% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cont.R
\name{check_cont}
\alias{check_cont}
\title{Check containers exist}
\usage{
check_cont(
  cont,
  domain = "https://hub.docker.com/v2/repositories/",
  verbose = TRUE
)
}
\arguments{
\item{cont}{Which Docker container to use on each OS
(\code{NULL} means no container will be used for that OS).
See 
\href{https://hub.docker.com/r/bioconductor/bioconductor_docker/tags}{here}
 for a list of all official Bioconductor Docker container versions.}

\item{domain}{Container repository API domain.}

\item{verbose}{Print messages.}
}
\value{
NUll
}
\description{
Check that a list of containers are valid.
}
\keyword{internal}
