% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_miniconda_args.R
\name{check_miniconda_args}
\alias{check_miniconda_args}
\title{Check Miniconda arguments}
\usage{
check_miniconda_args(
  runners,
  miniforge_variant,
  miniforge_version,
  activate_environment,
  environment_file,
  channels,
  verbose = TRUE
)
}
\arguments{
\item{runners}{Runner configurations for multiple Operating Systems (OS), 
including R versions, Bioc versions, and container sources.
Can use the \link[rworkflows]{construct_runners} functions to assist 
in constructing customized runners configurations.}

\item{miniforge_variant}{If provided, this variant of Miniforge will be
downloaded and installed. If \code{miniforge_variant=false}, 
Miniforge will not be installed at all.
If \code{miniforge_variant=""}, the "Miniforge3" variant will be installed.
If \code{miniforge_version} is not provided, the \code{latest} 
version will be used.
Currently-known values: - "Miniforge3" (default) - "Miniforge-pypy3" -
"Mambaforge" - "Mambaforge-pypy3".
Visit
https://github.com/conda-forge/miniforge/releases/ for more information on
available variants.}

\item{miniforge_version}{If provided, this version of the given Miniforge
variant will be downloaded and installed. If \code{miniforge_variant} is 
not provided, "Miniforge3" will be used. Visit
https://github.com/conda-forge/miniforge/releases/ for more information on
available versions.}

\item{activate_environment}{Environment name (or path) to activate on all 
shells. Default is "test" which will be created in \emph{$CONDA/envs/test}. 
If an empty string is used, no environment is activated by default 
(For "base" activation see the \code{auto-activate-base} option). 
If the environment does not exist, it will
be created and activated. If \code{environment-file} is used and you want 
that to be the environment used, you need to explicitly provide the name of
that environment on \code{activate-environment}. 
If using \code{sh/bash/cmd.exe}
shells please read the IMPORTANT! section on the README.md! to properly
activate conda environments on these shells.}

\item{environment_file}{Path or URL to a .yml file to build the 
conda environment with. For more information see
\href{https://docs.conda.io/projects/conda/en/latest/user-guide/tasks/manage-environments.html#creating-an-environment-from-an-environment-yml-file}{
here}.}

\item{channels}{Conda configuration. Comma separated list of channels 
to use in order of priority. See
\href{https://docs.conda.io/projects/conda/en/latest/user-guide/configuration/}{here}
for more information.}

\item{verbose}{Print messages.}
}
\value{
Named list of updated args
}
\description{
Check Miniconda arguments
}
\keyword{internal}
