% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_registry.R
\name{check_registry}
\alias{check_registry}
\title{Check registry}
\usage{
check_registry(registry)
}
\arguments{
\item{registry}{A character vector of registry names.}
}
\value{
A single registry name.
}
\description{
Check that the registry is valid.
}
\keyword{internal}
