% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda_export.R
\name{conda_export}
\alias{conda_export}
\title{Conda export}
\source{
https://stackoverflow.com/a/55687210
}
\usage{
conda_export(
  name,
  save_path = tempfile(fileext = "_requirements.txt"),
  preview = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{name}{Name of conda environment.}

\item{save_path}{Path to save the requirements file.}

\item{preview}{Print the requirements file to the R console.}

\item{verbose}{Print messages.}
}
\description{
Get a list of installed packages within a conda environment.
 Generates a requirements.txt file.
}
\examples{
\dontrun{
conda_export()
}
}
