% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies_latex.R
\name{dependencies_latex}
\alias{dependencies_latex}
\title{Latex dependencies}
\usage{
dependencies_latex(
  extra = c("bera", "nowidow", "parnotes", "marginfix", "etoolbox", "titlesec",
    "sectsty", "framed", "enumitem", "parskip", "soul", "placeins", "footmisc",
    "changepage", "xstring", "caption", "mathtools", "fancyhdr", "preprint", "ragged2e",
    "pdfcrop", "titling", "silence", "everysel"),
  install = TRUE,
  ...
)
}
\arguments{
\item{extra}{Additional dependencies to install.}

\item{install}{Install the packages with \link[tinytex]{tlmgr_install}
(\code{TRUE}). Otherwise, return a character vector of the packages to
install.}

\item{...}{
  Arguments passed on to \code{\link[tinytex:tlmgr]{tinytex::tlmgr_install}}
  \describe{
    \item{\code{usermode}}{(For expert users only) Whether to use TeX Live's
\href{https://www.tug.org/texlive/doc/tlmgr.html#USER-MODE}{user mode}. If
\code{TRUE}, you must have run \code{tlmgr('init-usertree')} once before.
This option allows you to manage a user-level texmf tree, e.g., install a
LaTeX package to your home directory instead of the system directory, to
which you do not have write permission. This option should not be needed on
personal computers, and has some limitations, so please read the
\pkg{tlmgr} manual very carefully before using it.}
    \item{\code{pkgs}}{A character vector of LaTeX package names.}
    \item{\code{path}}{Whether to run \code{tlmgr_path('add')} after installing packages
(\code{path = TRUE} is a conservative default: it is only necessary to do
this after a binary package is installed, such as the \pkg{metafont}
package, which contains the executable \command{mf}, but it does not hurt
even if no binary packages were installed).}
  }}
}
\value{
A character vector of extra latex packages.
}
\description{
Install extra latex package dependencies.
}
\examples{
dependencies_latex(install=FALSE)
}
