% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies_linux.R
\name{dependencies_linux}
\alias{dependencies_linux}
\title{Linux dependencies}
\usage{
dependencies_linux(
  os = "ubuntu",
  os_release = "20.04",
  extra = c("qpdf", "rsync", "git-core", "libicu-dev", "zlib1g-dev", "xfonts-100dpi",
    "xfonts-75dpi", "biber", "libsbml5-dev", "cmake", "pandoc", "pandoc-citeproc"),
  prefix = paste("apt-get update -y", "apt-get install -y", sep = " && "),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{os, os_release}{The operating system and operating system release version, see
\url{https://github.com/rstudio/r-system-requirements#operating-systems} for the
list of supported operating systems.

If \code{os_release} is \code{NULL}, \code{os} must consist of the operating system
and the version separated by a dash, e.g. \code{"ubuntu-18.04"}.}

\item{extra}{Additional dependencies to install.}

\item{prefix}{Prefix to prepend to the command.}

\item{verbose}{Print messages.}

\item{...}{
  Arguments passed on to \code{\link[remotes:system_requirements]{remotes::system_requirements}}
  \describe{
    \item{\code{os,os_release}}{The operating system and operating system release version, see
\url{https://github.com/rstudio/r-system-requirements#operating-systems} for the
list of supported operating systems.

If \code{os_release} is \code{NULL}, \code{os} must consist of the operating system
and the version separated by a dash, e.g. \code{"ubuntu-18.04"}.}
    \item{\code{path}}{The path to the dev package's root directory.}
    \item{\code{package}}{CRAN package name(s) to lookup system requirements for. If not
\code{NULL}, this is used and \code{path} is ignored.}
    \item{\code{curl}}{The location of the curl binary on your system.}
  }}
}
\value{
Command to install system dependencies.
}
\description{
Construct a command to install system dependencies for the 
Linux OS environment.
Includes all dependencies listed by \link[remotes]{system_requirements}, 
in addition to user-specified dependencies via \code{extra}.
}
\examples{
dependencies_linux(package="rworkflows")
}
