% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies_macos.R
\name{dependencies_macos}
\alias{dependencies_macos}
\title{MacOS dependencies}
\usage{
dependencies_macos(
  extra = c("libxml2", "harfbuzz", "fribidi", "libgit2", "texlive", "imagemagick@6",
    "rsync"),
  cask = c("xquartz"),
  prefix = "brew install",
  verbose = TRUE
)
}
\arguments{
\item{extra}{Additional dependencies to install.}

\item{cask}{Additional dependencies to install via cask.}

\item{prefix}{Prefix to prepend to the command.}

\item{verbose}{Print messages.}
}
\value{
Command to install system dependencies.
}
\description{
Construct a command to install system dependencies for the 
Mac OS environment.
}
\details{
### --- Justifications for each package --- ####
## libxml2 ###
Enable installing XML from source if needed
## imagemagick@6 ###
Required to install magick as noted at
https://github.com/r-lib/usethis/commit/f1f1e0d10c1ebc75fd4c18fa7e2de4551fd9978f#diff-9bfee71065492f63457918efcd912cf2
## harfbuzz fribidi ###
For textshaping, required by ragg, and required by pkgdown
## libgit2 ###
For installing usethis's dependency gert
## xquartz --cask ###
Required for tcltk
## libxfont ###
Required for some ggplot2 functions
## texlive ###
Required for rendering Sweave files (even with r-lib/actions/setup-tinytex)
}
\examples{
dependencies_macos()
}
