% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies_windows.R
\name{dependencies_windows}
\alias{dependencies_windows}
\title{Windows dependencies}
\usage{
dependencies_windows(extra = c("curl"), prefix = "npm install", verbose = TRUE)
}
\arguments{
\item{extra}{Additional dependencies to install.}

\item{prefix}{Prefix to prepend to the command.}

\item{verbose}{Print messages.}
}
\value{
Command to install system dependencies.
}
\description{
Construct a command to install system dependencies for the 
Windows environment.
}
\examples{
dependencies_windows()
}
