% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_authors.R
\name{get_authors}
\alias{get_authors}
\title{Get authors}
\usage{
get_authors(add_html = FALSE, file = ".")
}
\arguments{
\item{add_html}{Add HTML styling to certain fields (e.g "authors").}

\item{file}{DESCRIPTION file to use. By default the DESCRIPTION
file of the current package (i.e. the package the working directory
is part of) is used.}
}
\value{
Authors as a raw text or HTML string.
}
\description{
Get author names from the an R package DESCRIPTION file.
}
\keyword{internal}
