% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_description.R
\name{get_description}
\alias{get_description}
\title{Get DESCRIPTION}
\usage{
get_description(ref = NULL, path = here::here("DESCRIPTION"))
}
\arguments{
\item{ref}{Reference for a GitHub repository. If \code{NULL}
(the default), the reference is determined by the URL
field in the DESCRIPTION file.}

\item{path}{Path to DESCRIPTION file of an R package.}
}
\value{
\code{packageDescription} object
}
\description{
Find the DESCRIPTION file for a package
 through a variety of alternative methods.
}
\examples{
d <- get_description(ref="neurogenomics/rworkflows")
}
