% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_github_url_db.R
\name{get_github_url_db}
\alias{get_github_url_db}
\title{Get GitHub URL: from database}
\usage{
get_github_url_db(db, return_dt = FALSE)
}
\arguments{
\item{db}{Database.}

\item{return_dt}{Return data.table format.}
}
\value{
data.table or list
}
\description{
Extract the GitHub URLs from a tabular database of
 \emph{DESCRIPTION} file data.
}
\keyword{internal}
