% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_gha.R
\name{is_gha}
\alias{is_gha}
\title{Is GitHub Action}
\source{
\href{https://docs.github.com/en/actions/learn-github-actions/variables#default-environment-variables}{
GitHub Actions docs}
}
\usage{
is_gha(var = "GITHUB_ACTION", verbose = TRUE)
}
\arguments{
\item{var}{Environmental variable to check.}

\item{verbose}{Print messages.}
}
\value{
Boolean
}
\description{
Tests whether a function is currently being run within a GitHub Actions 
workflow or not.
}
\examples{
is_gha()
}
