% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_badges.R
\name{use_badges}
\alias{use_badges}
\title{Use badges}
\usage{
use_badges(
  ref = NULL,
  add_version = TRUE,
  add_actions = TRUE,
  add_commit = TRUE,
  add_codecov = TRUE,
  add_license = TRUE,
  add_authors = TRUE,
  action = "rworkflows",
  branch = "master",
  as_list = FALSE,
  sep = "\\n",
  verbose = TRUE
)
}
\arguments{
\item{ref}{Reference for a GitHub repository. If \code{NULL} (the default),
the reference is determined by the URL field in the DESCRIPTION file.}

\item{add_version}{Add package version with 
\link[badger]{badge_github_version}.}

\item{add_actions}{Add GitHub Actions status with
\link[badger]{badge_github_actions}.}

\item{add_commit}{Add the last GitHub repo commit date with
\link[badger]{badge_last_commit}.}

\item{add_codecov}{Add CodeCov status with
\link[badger]{badge_codecov}.}

\item{add_license}{Add license info with
\link[badger]{badge_license}.}

\item{add_authors}{Add author names inferred from 
the \code{DESCRIPTION} file.}

\item{action}{GitHub Actions workflow to create status badge for.}

\item{branch}{Name of the GitHub repository branch to use.}

\item{as_list}{Return the header as a named list (\code{TRUE}), 
or a collapsed text string (default: \code{FALSE}).}

\item{sep}{Character to separate each item in the list with 
using \link[base]{paste}.}

\item{verbose}{Print messages.}
}
\description{
Create one or more badges showing the status of your R package.
Uses the package 
\href{https://github.com/GuangchuangYu/badger}{\pkg{badger}}.
}
\examples{
## Causes issues bc examples can't find the  the DESCRIPTION file.
 \dontrun{
 rworkflows::use_badges() 
 }
}
