% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_gha.R
\name{use_gha}
\alias{use_gha}
\title{Use GitHub Actions}
\usage{
use_gha(
  name = "test-document-deploy",
  on = c("push", "pull_request"),
  branches = c("master", "main"),
  run_bioccheck = FALSE,
  run_crancheck = TRUE,
  has_testthat = TRUE,
  run_covr = TRUE,
  run_pkgdown = TRUE,
  has_runit = TRUE,
  run_docker = TRUE,
  repository = "${{ github.repository }}",
  github_token = "${{ secrets.PAT_GITHUB }}",
  docker_username = NULL,
  docker_org = docker_username,
  docker_token = "${{ secrets.DOCKER_TOKEN }}",
  cache_version = "cache_version",
  save_dir = here::here(".github", "workflows"),
  return_path = TRUE,
  force_new = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{name}{Workflow name.}

\item{on}{GitHub trigger conditions.}

\item{branches}{GitHub trigger branches.}

\item{run_bioccheck}{Run Bioconductor checks. 
Must pass in order to continue workflow.}

\item{run_crancheck}{Run CRAN checks. 
Must pass in order to continue workflow.}

\item{has_testthat}{Run unit tests and report results.}

\item{run_covr}{Run code coverage tests and publish results to codecov.}

\item{run_pkgdown}{Knit the \emph{README.Rmd} (if available), 
build documentation website, and deploy to \emph{gh-pages} branch.}

\item{has_runit}{Run R Unit tests.}

\item{run_docker}{Whether to build and push a Docker container to DockerHub.}

\item{repository}{GitHub repository to be checked.}

\item{github_token}{Token for the repo. 
Can be passed in using {{ secrets.PAT_GITHUB }}.}

\item{docker_username}{DockerHub username.}

\item{docker_org}{DockerHub organization name. 
Is the same as \code{docker_username} by default.}

\item{docker_token}{DockerHub token.}

\item{cache_version}{Name of the cache sudirectory to be used
when reinstalling software in GHA.}

\item{save_dir}{Directory to save workflow to.}

\item{return_path}{Return the path to the saved \emph{yaml} workflow file
(default: \code{TRUE}), or return the \emph{yaml} object directly.}

\item{force_new}{If the GHA workflow yaml already exists, 
overwrite with new one (default: \code{FALSE}).}

\item{verbose}{Print messages.}
}
\value{
Path or yaml object.
}
\description{
Use a \href{https://github.com/features/actions}{GitHub Actions (GHA)} 
workflow from
 \href{https://github.com/neurogenomics/rworkflows}{rworkflows}.
}
\examples{
 
path <- use_gha(save_dir = file.path(tempdir(),".github","workflows"))
}
