% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_vignette_getstarted.R
\name{use_vignette_getstarted}
\alias{use_vignette_getstarted}
\title{Use vignette: Get started}
\usage{
use_vignette_getstarted(
  package = names(get_description()),
  title = "Get started",
  vignette_index_entry = package,
  save_dir = here::here(),
  path = file.path(save_dir, "vignettes", paste0(package, ".Rmd")),
  output = "rmarkdown::html_document",
  force_new = FALSE,
  preview = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{package}{R package name.}

\item{title}{Title of vignette.}

\item{vignette_index_entry}{Index entry of the vignette, 
which is used when creating the navigation bar in the \pkg{pkgdown} site.}

\item{save_dir}{Directory to save workflow to.}

\item{path}{Path to the file.}

\item{output}{Vignette output style. 
Defaults to \link[rmarkdown]{html_document}.}

\item{force_new}{If the GHA workflow yaml already exists, 
overwrite with new one (default: \code{FALSE}).}

\item{preview}{Print the file to the R console.}

\item{verbose}{Print messages.}
}
\value{
Path to vignette file.
}
\description{
Creates a "Get started" rmarkdown vignette file.
}
\examples{
path <- use_vignette_getstarted(package = "mypackage",
                                ## use default save_dir in practice
                                save_dir = tempdir())
}
