% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda_export.R
\name{conda_export}
\alias{conda_export}
\title{Conda export}
\source{
https://stackoverflow.com/a/55687210
}
\usage{
conda_export(
  name,
  save_path = tempfile(fileext = "_requirements.txt"),
  preview = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{name}{Name of conda environment.}

\item{save_path}{Path to save the requirements file. 
If the file ends with .yml or .yaml, a conda-style 
yaml file will be generated.
If the file ends with requirements.txt, a pip-style requirements.txt file
 will be generated.}

\item{preview}{Print the requirements file to the R console.}

\item{verbose}{Print messages.}

\item{...}{Optional arguments, reserved for future expansion.}
}
\value{
Path to requirements file.
}
\description{
Get a list of installed packages within a conda environment.
 Generates a requirements.txt file.
}
\examples{
\dontrun{
conda_export()
}
}
