% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_authors.R
\name{get_authors}
\alias{get_authors}
\title{Get authors}
\usage{
get_authors(ref = NULL, add_html = FALSE)
}
\arguments{
\item{ref}{Reference for a GitHub repository. If \code{NULL}
(the default), the reference is determined by the URL
field in the DESCRIPTION file.}

\item{add_html}{Add HTML styling to certain fields (e.g "authors").}
}
\value{
Authors as a raw text or HTML string.
}
\description{
Get author names from the an R package DESCRIPTION file.
}
\keyword{internal}
