% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_readme.R
\name{use_readme}
\alias{use_readme}
\title{Use README}
\usage{
use_readme(
  save_dir = here::here(),
  path = file.path(save_dir, "README.Rmd"),
  force_new = FALSE,
  show = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{save_dir}{Directory to save the file to.}

\item{path}{Path to the file.}

\item{force_new}{If the file already exists, overwrite it 
(default: \code{FALSE}).}

\item{show}{Print the contents of the file in the R console.}

\item{verbose}{Print messages.}
}
\value{
Path to README file.
}
\description{
Creates an rmarkdown README file that autofills using metadata from the 
R package \emph{DESCRIPTION} file.
}
\examples{
## use default save_dir in practice
path <- use_readme(save_dir = tempdir()) 
}
